requirements.txt
discord.py[voice]
discord-py-interactions==3.*
youtube-dl